- added an env var that allows to see the background process console. - renamed servers.py to site_communication.py
